requirement.txt
ai21==3.0.1
ai21-tokenizer==0.12.0
aider-chat==0.72.3
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
async-timeout==4.0.3
atlassian-python-api==3.41.19
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chainlit==0.1.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
cryptography==44.0.1
dataclasses-json==0.6.7
datasets==3.3.2
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
faiss-gpu==1.7.2
fastapi==0.115.8
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grep-ast==0.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.17
langchain-ai21==1.0.1
langchain-community==0.3.16
langchain-core==0.3.37
langchain-experimental==0.3.4
langchain-huggingface==0.1.2
langchain-openai==0.3.6
langchain-text-splitters==0.3.5
langsmith==0.3.5
litellm==1.58.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.3.0
narwhals==1.25.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
ollama==0.4.7
openai==1.59.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdfminer.six==20240706
pexpect==4.9.0
pillow==10.4.0
portalocker==3.1.1
posthog==3.8.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pyarrow==19.0.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pypandoc==1.15
pypdf==5.3.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
ragas==0.2.13
referencing==0.36.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
streamlit==1.41.1
streamlit-js-eval==0.1.7
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
transformers==4.44.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
triton==3.1.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
